+++ tsconfig.json
@@ -14,5 +14,5 @@
     "removeComments": false,
     "preserveConstEnums": true
   },
-  "include": ["src/**/*", "src/components/hooks/appData.tsx", "src/components/.Dropdown.tsx", "src/components/.send.tsx", "src/components/.users.tsx", "src/components/.topics.tsx", "src/components/.seo.tsx", "src/components/.groups.tsx", "src/components/.add.tsx", "src/components/.image.tsx"]
+  "include": ["src/**/*", "src/components/hooks/appData.tsx", "src/components/.Dropdown.tsx", "src/components/.send.tsx", "src/components/.users.tsx", "src/components/.topics.tsx", "src/components/.seo.tsx", "src/components/.groups.tsx", "src/components/.add.tsx", "src/components/.image.tsx", "src/components/hooks/.test.tsx"]
 }
